move folders, remove singleton
