+++ lib/services/permission/check_permission_service.dart
@@ -1,11 +1,6 @@
 import 'package:permission_handler/permission_handler.dart';
 
 class CheckPermissionService {
-  static final CheckPermissionService _instance =
-      CheckPermissionService._internal();
-  factory CheckPermissionService() => _instance;
-  CheckPermissionService._internal();
-
   Future<void> checkPermission(Permission permission) async {
     final status = await permission.status;
     if (!status.isGranted) {
