+++ lib/ui/home/home_view.dart
@@ -4,7 +4,7 @@ import 'package:medication_tracker/data/providers/medication_provider.dart';
 import 'package:medication_tracker/data/providers/profile_provider.dart';
 import 'package:medication_tracker/ui/edit_medication/edit_medication_view.dart';
 import 'package:medication_tracker/ui/search/fda_search_view.dart';
-import 'package:medication_tracker/ui/select_profile_view.dart';
+import 'package:medication_tracker/ui/select_profile/select_profile_view.dart';
 import 'package:medication_tracker/ui/home/med_tile.dart';
 import 'package:medication_tracker/ui/home/nav_bar.dart';
 import 'package:provider/provider.dart';
