+++ lib/ui/search/fda_search_view.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:medication_tracker/data/providers/fda_api_provider.dart';
 import 'package:medication_tracker/services/image/image_service.dart';
 import 'package:medication_tracker/ui/core/black_button.dart';
-import 'package:medication_tracker/ui/create_medication_view.dart';
+import 'package:medication_tracker/ui/create_medication/create_medication_view.dart';
 import 'package:medication_tracker/ui/core/header.dart';
 import 'package:medication_tracker/ui/core/photo_upload_row.dart';
 import 'package:provider/provider.dart';
