+++ lib/ui/select_profile/select_profile_view.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:medication_tracker/data/providers/profile_provider.dart';
-import 'package:medication_tracker/ui/create_profile_view.dart';
+import 'package:medication_tracker/ui/create_profile/create_profile_view.dart';
 import 'package:medication_tracker/ui/home/home_view.dart';
 import 'package:medication_tracker/ui/core/black_button.dart';
 import 'package:medication_tracker/ui/core/header.dart';
